+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 15-12-2025
+
+- Remove `IERC7802`, `IERC7821` and `IERC7786`. These interfaces were migrated to `@openzeppelin/contracts>=5.5.0`.
+- Remove `ERC7786Receiver`. This contract was migrated to `@openzeppelin/contracts>=5.5.0`.
+- Updated crosschain bridge contracts terminology from "Recipient" to "Receiver" to align with ERC-7786 specification.
+- `PaymasterCore`: Update default version of the ERC-4337 entrypoint to v0.9.
+
 ## 02-11-2025
 
 - Remove `WebAuthn`, `SignerWebAuthn`, `ERC7913WebAuthnVerifier`. These contracts were migrated to `@openzeppelin/contracts>=5.5.0`.